requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
bcrypt==4.2.0
cffi==1.17.1
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.4.0
Fabric3==1.14.post1
fonttools==4.54.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
llvmlite==0.41.1
matplotlib==3.7.5
numba==0.58.1
numpy==1.24.4
packaging==24.2
paramiko==2.12.0
pillow==10.4.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
xcover==0.2.4
zipp==3.20.2
